Implement multi-gpu training in HPO with Ray Tune and Ray Train (#277)
* implement multi-gpu training in HPO with Ray Tune and Ray Train

* update flatiron scripts for HPO in pytorch

* update flatiron scripts for hpo on pytorch backend

* simplify ray hpo training function

* implement distributed training using Ray Train

* add support for HyperOpt and ASHA

* clean up commented code

* feat: distributed multi-node training using Ray Train

* fix CI test

* feat: enable comet ml logging when using Ray Train
