+++ parameters/pyg-cms-small.yaml
@@ -5,12 +5,12 @@ data_dir:
 gpus: "0"
 gpu_batch_multiplier: 1
 load:
-num_epochs: 2
+num_epochs: 10
 patience: 20
 lr: 0.0001
-conv_type: gnn_lsh
-ntrain:
-ntest:
+conv_type: gravnet
+ntrain: 500
+ntest: 500
 nvalid: 500
 num_workers: 0
 prefetch_factor:
@@ -43,6 +43,26 @@ model:
     width: 256
     num_convs: 3
     dropout: 0.0
+
+raytune:
+  local_dir: # Note: please specify an absolute path
+  sched: asha # asha, hyperband
+  search_alg: hyperopt # bayes, bohb, hyperopt, nevergrad, scikit
+  default_metric: "val_loss"
+  default_mode: "min"
+  # Tune schedule specific parameters
+  asha:
+    max_t: 200
+    reduction_factor: 4
+    brackets: 1
+    grace_period: 10
+  hyperband:
+    max_t: 200
+    reduction_factor: 4
+  hyperopt:
+    n_random_steps: 10
+  nevergrad:
+    n_random_steps: 10
 
 train_dataset:
   cms:
