+++ scripts/flatiron/pt_train.slurm
@@ -1,12 +1,13 @@
 #!/bin/sh
 
 # Walltime limit
-#SBATCH -t 2:00:00
+#SBATCH -t 168:00:00
 #SBATCH -N 1
 #SBATCH --exclusive
 #SBATCH --tasks-per-node=1
 #SBATCH -p gpu
 #SBATCH --gpus-per-task=4
+#SBATCH --cpus-per-task=64
 #SBATCH --constraint=h100,ib
 
 # Job name
@@ -36,12 +37,9 @@ CUDA_VISIBLE_DEVICES=0,1,2,3,4,5,6,7 python3 -u mlpf/pyg_pipeline.py --train \
     --config $1 \
     --prefix $2 \
     --gpus "0,1,2,3,4,5,6,7" \
-    --lr 0.0001 \
     --gpu-batch-multiplier 4 \
     --num-workers 1 \
     --prefetch-factor 2 \
-    --checkpoint-freq 1 \
-    --comet \
-    --comet-step-freq 20
+    --checkpoint-freq 1
 
 echo 'Training done.'
